pyproject.toml
[project]
name = "ai-career-coach"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.7",
    "langgraph>=0.6.5",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.48.1",
    "uvicorn>=0.35.0",
]
